Projects/Youtube Downloader/requirements.txt
Pillow==7.2.0
pytube3==9.6.4

